v2

+++ app/src/main/java/com/example/nearstore/StoreModal.kt
@@ -1,6 +1,6 @@
 package com.example.nearstore
 
-data class StoreModal(var storeid : Int = 0,var storename : String = "", var storelocation : String = " ", var storerating : Int = 0, val storelat : Int = 0, val storelong : Int = 0, ){
+data class StoreModal(var storeid : Int = 0,var storename : String = "", var storelocation : String = " ", var storerating : Int = 0, val storelat : Double = 0.0, val storelong : Double = 0.0, ){
 
 
 }
